+++ src/mine/ui/controls/api/include/mine/ui/controls/ControlsAll.h
@@ -10,6 +10,7 @@
 #include <mine/ui/controls/ContentPresenter.h>
 #include <mine/ui/controls/ContentControl.h>
 #include <mine/ui/controls/UserControl.h>
+#include <mine/ui/controls/Page.h>
 #include <mine/ui/controls/TextBlock.h>
 #include <mine/ui/controls/Button.h>
 #include <mine/ui/controls/Border.h>